update web-vault to v2023.8.2 and rename project
in web-v2023.8.x the pageTitle (i.e. "$APP_NAME Web Vault") was removed,
so instead of using the hardcoded replacement, we replace the name of
the project `Vaultwarden Web`

the footer is changed again from v2023.7.1 for better legibility.

compared to the previous patch files some changes are purely cosmetic due
to running prettier and eslint.

and we don't need to change or pack the 404.html file since Vaultwarden
already provides a 404 page